feat(ui-bulk): Create common points component, add javadocs
